refactor: replace sliden with slide core
