+++ src/Packages/cn.org.vrcd.vpm.aaas.slide/Scripts/Controller/SlidePresenterRemoteController.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 32019d2972d7c3a4c979b7b76bad83d0
+guid: c2f23f1197d158342b4b4fe4f179c1ce
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
