+++ src/Packages/cn.org.vrcd.vpm.aaas.slide/Scripts/Controller/SlidePresenterRemoteController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7357f4fb45984ffc9872719f6b4c34aa
+timeCreated: 1753459086
