--- src/Packages/cn.org.vrcd.vpm.aaas/Scripts/Other/Telecommand.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b845251a0ac5aea4b8b8a6391c0cc913
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
