--- src/Packages/cn.org.vrcd.vpm.aaas/Scripts/Other/Telecommand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6ba6ead063a3a5447bf53b00a8eafffb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
