+++ src/Packages/cn.org.vrcd.vpm.aaas/Textures.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 68af019af2ee4ef47b22dc5008df35e9
-AssemblyDefinitionImporter:
+guid: 5cca6da946f33d142b2ab75b3b74d49c
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
