+++ src/Packages/cn.org.vrcd.vpm.aaas/Textures/AAAS-SlideVideoTexture.renderTexture.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: f0d5e06008e26c648be421ac2557517a
+guid: 90737f3092063454cb4a25ab90480362
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
+  mainObjectFileID: 8400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
