--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5d505c50b3c01be488e808432fb55195
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
