--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fbae424f39d11524b8033e850e28e1c0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
