--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/ArrayUtils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b9c4482a08d53d4a9816791a4447096
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
