--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/ChikuwaProducts.Sliden.asmdef
@@ -1,4 +1,0 @@
-{
-	"name": "ChikuwaProducts.Sliden",
-	"references":[ "GUID:99835874ee819da44948776e0df4ff1d", "GUID:3c1bc1267eab5884ebe7f232c09ee0d9" ]
-}
