--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/ChikuwaProducts.Sliden.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 64e3b1e53fb25c04981f7a631b7728ce
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
