--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Hidable.cs
@@ -1,20 +1,0 @@
-﻿
-using UdonSharp;
-using UnityEngine;
-using VRC.SDKBase;
-using VRC.Udon;
-
-namespace cdes_presets
-{
-    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
-    public class Hidable : UdonSharpBehaviour
-    {
-        public Sliden Sliden;
-        void Start()
-        {
-            if (Sliden) {
-                Sliden.AddHidable(gameObject);
-            }
-        }
-    }
-}
