--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/LocalTablet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9fe1b3f1645a53740897f285df02cf82
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
