--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/LocalTablet.cs
@@ -1,16 +1,0 @@
-﻿using UdonSharp;
-
-namespace cdes_presets
-{
-
-    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
-    public class LocalTablet : Tablet
-    {
-        protected override void Start()
-        {
-            base.Start();
-            Lock = false;
-        }
-    }
-
-}
