--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/LocalTablet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8fabee1d3cf2fcd4e8b37fa66f7d10ac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
