--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/OffSet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4a8a8b71f5a2c944d9656342a915aa91
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
