--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/OffSet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0ec9e3dfbc399e4ebdb618cf0824a28
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
