--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Screen.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ccbec60f284cc0f4ea04757de99eb359
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
