--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Screen.cs
@@ -1,19 +1,0 @@
-﻿
-using UdonSharp;
-
-namespace cdes_presets
-{
-
-    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
-    public class Screen : UdonSharpBehaviour
-    {
-        public Sliden Sliden;
-        void Start()
-        {
-            if (Sliden) {
-                Sliden.AddScreen(gameObject);
-            }
-        }
-    }
-}
-
