--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Sliden.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 32a71bd09c6b00f44af1ec902373e1b6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
