--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/SlidenListener.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7741ccee7762f1e41a26611ca02d478e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
