--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/SlidenListener.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 798c540a0c8041c48aa452b0e5845dca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
