--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/SpeakerTablet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ecef54cd7d61b224cb0835ccd9358307
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
