--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/SpeakerTablet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4161b6e47c3a16a4ab60d2981617f42a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
