--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Tablet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 673e27849c4bb7849b5aeadd5a0b48bd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
