--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/Tablet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6bb9fc6d4b83913499be5c219b211154
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
