--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/TabletPlaceholder.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bc91ccffbdd42504484e53f954f8e65a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
