--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/TabletPlaceholder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a7ee2dc19432c2a46853dc154da595f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
