--- src/Packages/cn.org.vrcd.vpm.aaas/Third-part/ChikuwaProducts/Sliden/ToggleButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e0257c5f6b2a0ae49b2aadd82f396504
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
